moving zip to module
